html: add floating magnifier showing details in the result image (mumax#216)
* Adds a floating magnifier to show details in the result image

This adds a magnifier to the html interface of the mumax results page, that shows a zoomed in detail of the result image.

* fixed loupe for very small images

In very small images (and near the edges of all images) the loupe wasn't updated.

* images shows sharp pixels

The loupe/zoomed image conserves the sharp pixels (no smoothing, no interpolation) which corresponds better to the simulation cells.