Add FreeLayerThickness parameter
The Slonczewski torque depends on the thickness of the free layer.
Before this commit, this value was deduced from the thickness of the
mesh. There could be cases in which this is not what the user wants.

The changes made in this commit allow the user to set the free layer
thickness by hand through the "FreeLayerThickness" regionwise parameter.

If FreeLayerThickness is set to zero (the default), then the thickness
will be deduced from the thickness of the mesh, ensuring backwards
compatibility.